consolidated network files
