+++ packages/houdini/src/generators/runtime/template/index.ts
@@ -3,9 +3,6 @@ import { GraphQLTagResult } from './types'
 export * from './network'
 export * from './types'
 
-// @ts-ignore: file doesn't exist yet
-export * from './environment'
-
 export { default as query, getQuery } from './query'
 export { default as mutation } from './mutation'
 export { default as fragment } from './fragment'
