+++ packages/houdini/src/generators/runtime/template/network.ts
@@ -1,6 +1,3 @@
-// @ts-ignore: file doesn't exist yet
-import type { Environment } from './environment'
-
 export type FetchParams = {
 	text: string
 	variables: { [key: string]: any }
@@ -96,4 +93,19 @@ export type SubscriptionHandler = {
 			complete: () => void
 		}
 	) => () => void
+}
+
+export class Environment {
+	private fetch: RequestHandler
+	socket: SubscriptionHandler | null | undefined
+
+	// this project uses subscriptions so make sure one is passed when constructing an environment
+	constructor(networkFn: RequestHandler, subscriptionHandler?: SubscriptionHandler | null) {
+		this.fetch = networkFn
+		this.socket = subscriptionHandler
+	}
+
+	sendRequest(ctx: FetchContext, params: FetchParams, session?: FetchSession) {
+		return this.fetch.call(ctx, params, session)
+	}
 }
